FPS 3D game
Game Video:
https://www.youtube.com/watch?v=79vduucOAjw

Game description:
The player must kill every zombie in order to rescue his farm animal, and zombie NPCs have to follow the player in order to kill him. The next step is for the user to tag each animal so that it will follow him into the farm barn.

I used navmesh and ai pathfinding for the zombies to follow the player.